+++ src/includepriv/xmmspriv/xmms_ipc.h
@@ -15,8 +15,8 @@ typedef struct xmms_ipc_manager_St xmms_ipc_manager_t;
 xmms_ipc_manager_t *xmms_ipc_manager_get (void);
 
 gboolean xmms_ipc_has_pending (guint signalid);
-void xmms_ipc_send_message (gint64 cli, xmms_ipc_msg_t *msg, xmms_error_t *err);
-void xmms_ipc_send_broadcast (guint broadcastid, gint64 cli, xmmsv_t *arg, xmms_error_t *err);
+void xmms_ipc_send_message (gint cli, xmms_ipc_msg_t *msg, xmms_error_t *err);
+void xmms_ipc_send_broadcast (guint broadcastid, gint cli, xmmsv_t *arg, xmms_error_t *err);
 GList *xmms_ipc_get_connected_clients (void);
 
 #endif
